Renamed StringExtensions.cs to Extensions.cs Moved ReplaceDot method to Extensions.cs as an extension method Renamed ReplaceDot as ReplaceDotInKeys
