+++ src/NLog.Targets.ElasticSearch/NLog.Targets.ElasticSearch.csproj
@@ -51,7 +51,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="IElasticSearchTarget.cs" />
-    <Compile Include="StringExtensions.cs" />
+    <Compile Include="Extensions.cs" />
     <Compile Include="Field.cs" />
     <Compile Include="ElasticSearchTarget.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
